backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
python-dotenv==1.0.0
httpx>=0.28.1
pydantic[email]>=2.12.5
clerk-backend-api>=4.2.0
python-multipart>=0.0.21
